Replace LocationRepository with Intent-service
